Change account common feature to theme 2
