+++ feature/account/common/src/debug/kotlin/app/k9mail/feature/account/common/ui/AppTitleTopHeaderPreview.kt
@@ -1,0 +1,13 @@
+package app.k9mail.feature.account.common.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import app.k9mail.core.ui.compose.designsystem.PreviewWithThemes
+
+@Composable
+@Preview(showBackground = true)
+internal fun AppTitleTopHeaderPreview() {
+    PreviewWithThemes {
+        AppTitleTopHeader(title = "Title")
+    }
+}
