+++ feature/account/common/src/main/kotlin/app/k9mail/feature/account/common/ui/AccountTopAppBar.kt
@@ -2,9 +2,7 @@ package app.k9mail.feature.account.common.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import app.k9mail.core.ui.compose.common.annotation.PreviewDevices
 import app.k9mail.core.ui.compose.designsystem.organism.TopAppBar
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 
 /**
  * Top app bar for the account screens.
@@ -18,14 +16,4 @@ fun AccountTopAppBar(
         title = title,
         modifier = modifier,
     )
-}
-
-@PreviewDevices
-@Composable
-internal fun AccountTopAppBarPreview() {
-    PreviewWithThemes {
-        AccountTopAppBar(
-            title = "Title",
-        )
-    }
 }
