+++ feature/account/common/src/main/kotlin/app/k9mail/feature/account/common/ui/AppTitleTopHeader.kt
@@ -11,12 +11,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import app.k9mail.core.ui.compose.designsystem.atom.text.TextDisplayMedium
 import app.k9mail.core.ui.compose.designsystem.template.ResponsiveWidthContainer
-import app.k9mail.core.ui.compose.theme.MainTheme
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
+import app.k9mail.core.ui.compose.theme2.MainTheme
 import app.k9mail.feature.account.common.R
 
 private const val TITLE_ICON_SIZE_DP = 56
@@ -57,13 +55,5 @@ fun AppTitleTopHeader(
 
             TextDisplayMedium(text = title)
         }
-    }
-}
-
-@Preview
-@Composable
-internal fun AppTitleTopHeaderPreview() {
-    PreviewWithThemes {
-        AppTitleTopHeader(title = "Title")
     }
 }
