+++ feature/account/common/src/main/kotlin/app/k9mail/feature/account/common/ui/ContentListView.kt
@@ -11,11 +11,8 @@ import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextTitleMedium
 import app.k9mail.core.ui.compose.designsystem.template.ResponsiveWidthContainer
-import app.k9mail.core.ui.compose.theme.MainTheme
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
+import app.k9mail.core.ui.compose.theme2.MainTheme
 
 @Composable
 fun ContentListView(
@@ -39,24 +36,6 @@ fun ContentListView(
             verticalArrangement = verticalArrangement,
         ) {
             items()
-        }
-    }
-}
-
-@Preview
-@Composable
-internal fun ContentListViewPreview() {
-    PreviewWithThemes {
-        ContentListView {
-            item {
-                TextTitleMedium("Item 1")
-            }
-            item {
-                TextTitleMedium("Item 2")
-            }
-            item {
-                TextTitleMedium("Item 3")
-            }
         }
     }
 }
